services/bounds-calibration/pyproject.toml
# bounds-calibration — scheduled Cloud Run Job. Reads rpc_predictions
# percentiles, proposes new min/max bounds as a GitHub PR against runtime config.
[project]
name = "msm-bounds-calibration"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "google-cloud-bigquery>=3.25",
    "PyGithub>=2.3",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-cov>=5", "import-linter>=2.0"]

[tool.pytest.ini_options]
addopts = "--cov=msm_bounds --cov-fail-under=85"
testpaths = ["tests"]
